requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
streamlit>=1.28.0
pyyaml>=6.0.1

# Testing and monitoring
pytest>=7.4.0
pytest-cov>=4.1.0
psutil>=5.9.0
tqdm>=4.66.1

# HTTP and API
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
numpy>=1.24.0
pandas>=2.1.0

# Logging and monitoring
python-json-logger>=2.0.7
